:recycle: refactor(scraper): share cursor helpers between scrapers
The incremental-walk cursor logic was written for congress.gov but is not
federal in any way — any oldest-first walk with a per-item outcome needs
exactly this. Move it to utils/ so the next scraper reuses the reasoning
instead of reimplementing it, congress.ts re-exports both so its existing
tests keep pointing at the same names.
